+++ SprOutput.c
@@ -22,6 +22,9 @@
   CJB: 07-May-25: Dogfooding the _Optional qualifier.
 */
 
+/* ISO library headers */
+#include <stdint.h>
+
 /* Acorn C/C++ library headers */
 #include "kernel.h"
 #include "swis.h"
@@ -57,8 +60,8 @@ static _Optional _kernel_oserror *switch_output_common(SpriteAreaHeader *const a
   /* Find the buffer size required for the save area */
   _kernel_swi_regs regs;
   regs.r[0] = SPRITEOP_USERAREA_SPRNAME + SPRITEOP_READ_SAVE_AREA_SIZE;
-  regs.r[1] = (int)area;
-  regs.r[2] = name ? (int)name : 0;
+  regs.r[1] = (intptr_t)area;
+  regs.r[2] = name ? (intptr_t)name : 0;
 
   _Optional _kernel_oserror *e = _kernel_swi(OS_SpriteOp, &regs, &regs);
 
@@ -87,9 +90,9 @@ static _Optional _kernel_oserror *switch_output_common(SpriteAreaHeader *const a
         assert(buffer != NULL);
         regs.r[0] = reason + (area ? SPRITEOP_USERAREA_SPRNAME :
                                      SPRITEOP_SYSTEMAREA);
-        regs.r[1] = (int)area;
-        regs.r[2] = name ? (int)name : 0;
-        regs.r[3] = (int)buffer;
+        regs.r[1] = (intptr_t)area;
+        regs.r[2] = name ? (intptr_t)name : 0;
+        regs.r[3] = (intptr_t)buffer;
         e = _kernel_swi(OS_SpriteOp, &regs, &regs);
         if (e != NULL)
         {
