+++ SprRename.c
@@ -22,6 +22,9 @@
   CJB: 07-May-25: Dogfooding the _Optional qualifier.
 */
 
+/* ISO library headers */
+#include <stdint.h>
+
 /* Acorn C/C++ library headers */
 #include "kernel.h"
 #include "swis.h"
@@ -41,9 +44,9 @@ _Optional _kernel_oserror *os_sprite_op_rename(SpriteAreaHeader *const area,
 
   _kernel_swi_regs regs;
   regs.r[0] = SPRITEOP_USERAREA_SPRNAME + SPRITEOP_RENAME;
-  regs.r[1] = (int)area;
-  regs.r[2] = (int)old_name;
-  regs.r[3] = (int)new_name;
+  regs.r[1] = (intptr_t)area;
+  regs.r[2] = (intptr_t)old_name;
+  regs.r[3] = (intptr_t)new_name;
 
   return _kernel_swi(OS_SpriteOp, &regs, &regs);
 }
