+++ SprRestore.c
@@ -22,6 +22,9 @@
   CJB: 07-May-25: Dogfooding the _Optional qualifier.
 */
 
+/* ISO library headers */
+#include <stdint.h>
+
 /* Acorn C/C++ library headers */
 #include "kernel.h"
 #include "swis.h"
@@ -40,9 +43,9 @@ _Optional _kernel_oserror *os_sprite_op_restore_output(
 
   _kernel_swi_regs regs;
   regs.r[0] = output->reason;
-  regs.r[1] = (int)output->area;
-  regs.r[2] = (int)output->name_or_addr;
-  regs.r[3] = (int)output->buffer;
+  regs.r[1] = (intptr_t)output->area;
+  regs.r[2] = (intptr_t)output->name_or_addr;
+  regs.r[3] = (intptr_t)output->buffer;
 
   return _kernel_swi(OS_SpriteOp, &regs, &regs);
 }
