+++ SprSave.c
@@ -23,6 +23,9 @@
   CJB: 07-May-25: Dogfooding the _Optional qualifier.
 */
 
+/* ISO library headers */
+#include <stdint.h>
+
 /* Acorn C/C++ library headers */
 #include "kernel.h"
 #include "swis.h"
@@ -41,8 +44,8 @@ _Optional _kernel_oserror *os_sprite_op_save(SpriteAreaHeader *const area,
 
   _kernel_swi_regs regs;
   regs.r[0] = SPRITEOP_USERAREA_SPRNAME + SPRITEOP_SAVE_AREA;
-  regs.r[1] = (int)area;
-  regs.r[2] = (int)file_name;
+  regs.r[1] = (intptr_t)area;
+  regs.r[2] = (intptr_t)file_name;
 
   return _kernel_swi(OS_SpriteOp, &regs, &regs);
 }
