+++ SprSelect.c
@@ -23,6 +23,9 @@
   CJB: 07-May-25: Dogfooding the _Optional qualifier.
 */
 
+/* ISO library headers */
+#include <stdint.h>
+
 /* Acorn C/C++ library headers */
 #include "kernel.h"
 #include "swis.h"
@@ -41,8 +44,8 @@ _Optional _kernel_oserror *os_sprite_op_select(SpriteAreaHeader *const area,
 
   _kernel_swi_regs regs;
   regs.r[0] = SPRITEOP_USERAREA_SPRNAME + SPRITEOP_SELECT_SPRITE;
-  regs.r[1] = (int)area;
-  regs.r[2] = (int)name;
+  regs.r[1] = (intptr_t)area;
+  regs.r[2] = (intptr_t)name;
 
   _Optional _kernel_oserror *const e = _kernel_swi(OS_SpriteOp, &regs, &regs);
   if (e == NULL && area) {
